requirements.txt
numpy
jax
jaxlib
opencv-python
Pillow
tensorboard
tensorflow
gin-config
dm_pix
rawpy
mediapy
lpips_jax
chex
optax
ml-dtypes
flax
gdown
torch
torchvision
torchaudio
orbax-checkpoint==0.3.5
matplotlib==3.8.4